Modification in 'fourier_filter' and 'mask'
git-svn-id: file:///home/roberto/Git/svn@6086 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: f7836a70c1502f6ee412d0f0015a56e848d783e8

+++ trunk/xmipp/libraries/data/mask.h
@@ -276,7 +276,7 @@ void BinaryConeMask(MultidimArray< int >& mask,
  * outside (ie in the missing wedge) it will be zero. The mask is centered at
  * (x0,y0,z0), and rotated with respect to euler angle matrix A.
  */
-void BinaryWedgeMask(MultidimArray< double >& mask, double theta0, double thetaF,
+void BinaryWedgeMask(MultidimArray< int >& mask, double theta0, double thetaF,
                      Matrix2D< double > A);
